Move moat to its own module
