+++ .idea/compiler.xml
@@ -5,6 +5,7 @@
       <module name="sa-mobile-sdk-android.app" target="1.8" />
       <module name="sa-mobile-sdk-android.saadmob" target="1.8" />
       <module name="sa-mobile-sdk-android.saair" target="1.8" />
+      <module name="sa-mobile-sdk-android.samoat" target="1.8" />
       <module name="sa-mobile-sdk-android.samopub" target="1.8" />
       <module name="sa-mobile-sdk-android.saunity" target="1.8" />
       <module name="sa-mobile-sdk-android.superawesome" target="1.8" />
