+++ .idea/modules.xml
@@ -6,6 +6,7 @@
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/app/sa-mobile-sdk-android.app.iml" filepath="$PROJECT_DIR$/.idea/modules/app/sa-mobile-sdk-android.app.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/saadmob/sa-mobile-sdk-android.saadmob.iml" filepath="$PROJECT_DIR$/.idea/modules/saadmob/sa-mobile-sdk-android.saadmob.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/saair/sa-mobile-sdk-android.saair.iml" filepath="$PROJECT_DIR$/.idea/modules/saair/sa-mobile-sdk-android.saair.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/samoat/sa-mobile-sdk-android.samoat.iml" filepath="$PROJECT_DIR$/.idea/modules/samoat/sa-mobile-sdk-android.samoat.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/samopub/sa-mobile-sdk-android.samopub.iml" filepath="$PROJECT_DIR$/.idea/modules/samopub/sa-mobile-sdk-android.samopub.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/saunity/sa-mobile-sdk-android.saunity.iml" filepath="$PROJECT_DIR$/.idea/modules/saunity/sa-mobile-sdk-android.saunity.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/superawesome/sa-mobile-sdk-android.superawesome.iml" filepath="$PROJECT_DIR$/.idea/modules/superawesome/sa-mobile-sdk-android.superawesome.iml" />
