+++ app/build.gradle
@@ -50,7 +50,7 @@ dependencies {
     implementation('com.mopub:mopub-sdk-fullscreen:5.13.0') { transitive = true }
     implementation('com.mopub:mopub-sdk-native-static:5.13.0') { transitive = true }
 
-    implementation 'com.google.android.gms:play-services-ads:19.1.0'
+    implementation 'com.google.android.gms:play-services-ads:19.2.0'
     implementation "androidx.multidex:multidex:2.0.1"
     implementation 'androidx.appcompat:appcompat:1.1.0'
     implementation "androidx.core:core-ktx:1.3.0"
