+++ samoat/src/main/AndroidManifest.xml
@@ -1,0 +1,4 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="tv.superawesome.plugins.publisher.moat">
+
+</manifest>
