+++ superawesome/build.gradle
@@ -84,4 +84,5 @@ repositories {
 
 dependencies {
     api project(':superawesome-base')
+    implementation project(':samoat')
 }
